Servlets module added

+++ servlets/src/main/java/com/oddlabs/matchservlet/GetFirstIntQuery.java
@@ -4,8 +4,9 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 final strictfp class GetFirstIntQuery implements Query {
+
 	public final Object process(ResultSet result) throws SQLException {
 		result.first();
-		return (Integer)result.getInt(1);
+		return result.getInt(1);
 	}
 }
